icu_casemap

Module internals

This module contains most of the actual algorithms for case mapping.

Primarily, it implements methods on CaseMap, which contains the data model.

Structs

ContextIterator 🔒

FoldOptions 🔒

FullCaseWriteable 🔒

StringAndWriteable 🔒

Helper type that wraps a writeable in a prefix string

Enums

CaseMapLocale

FullMappingResult

Constants

ACUTE 🔒