Struct icu_casemap::internals::ContextIterator

pub(crate) struct ContextIterator<'a> {
    before: &'a str,
    after: &'a str,
}

Fields
before: &'a strafter: &'a str
Implementations

impl<'a> ContextIterator<'a>

pub fn new(before: &'a str, char_and_after: &'a str) -> Self

fn add_greek_diacritics(&self, diacritics: GreekDiacritics) -> GreekDiacritics

fn preceded_by_greek_letter(&self) -> bool

fn preceding_greek_vowel_diacritics( &self, ) -> Option<GreekCombiningCharacterSequenceDiacritics>

fn preceded_by_soft_dotted(&self, mapping: &CaseMapV1<'_>) -> bool

fn preceded_by_capital_i<const I_MUST_NOT_START_STRING: bool>( &self, mapping: &CaseMapV1<'_>, ) -> bool

Checks if the preceding character is a capital I, allowing for non-Above combining characters in between.

If I_MUST_NOT_START_STRING is true, additionally will require that the capital I does not start the string

fn preceded_by_cased_letter(&self, mapping: &CaseMapV1<'_>) -> bool

fn followed_by_cased_letter(&self, mapping: &CaseMapV1<'_>) -> bool

fn followed_by_more_above(&self, mapping: &CaseMapV1<'_>) -> bool

fn followed_by_dot_above(&self, mapping: &CaseMapV1<'_>) -> bool

fn is_dutch_ij_pair_at_beginning(&self, mapping: &CaseMapV1<'_>) -> bool

Checks the preceding and surrounding context of a j or J and returns true if it is preceded by an i or I at the start of the string. If one has an acute accent, both must have the accent for this to return true. No other accents are handled.