icu_casemap

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Modules

data

The primary per-codepoint casefolding data

exception_helpers

This module contains various types for the header part of casemapping exception data

exceptions

This is the main module pertaining to casemapping exceptions.

Structs

Baked

Baked data

CaseMap

This type contains all of the casemapping data

CaseMapUnfold

Reverse case folding data. Maps from multi-character strings back to code-points that fold to those strings.

CaseMapUnfoldV1

Reverse case mapping data.

CaseMapV1

Marker for casemapping data.