Module icu_casemap::titlecase

Titlecasing-specific try_new_with_mapper_unstable

Structs

TitlecaseMapper

A wrapper around CaseMapper that can compute titlecasing stuff, and is able to load additional data to support the non-legacy “head adjustment” behavior.

TitlecaseOptions

Various options for controlling titlecasing

Enums

LeadingAdjustment

Where to start casing the string.

TrailingCase

How to handle the rest of the string once the beginning of the string has been titlecased.