icu_casemap::provider
Module exceptions

This is the main module pertaining to casemapping exceptions.

A single exception is represented by the Exception type and its ULE equivalent.

The storage format is complicated (and documented on Exception), but the data format is represented equally by DecodedException, which is more human-readable.

Structs

CaseMapExceptions

This represents case mapping exceptions that can’t be represented as a delta applied to the original code point. The codepoint trie in CaseMapper stores indices into this VarZeroVec.

DecodedException

A decoded Exception type, with all of the data parsed out into separate fields.

Exception

A type representing the wire format of Exception. The data contained is equivalently represented by DecodedException.

ExceptionULE

VarULE type for Exception. See Exception for documentation.